requirements.txt
pyserial==3.5b0
schedule==1.1.0
slack-sdk==3.18.3
slack-bolt==1.15.0
gsm0338==1.0.0
jinja2==2.11.3

autopep8
flake8